falta hacer constructor vacío de Room (Creo)

+++ src/main/java/com/esolutions/trainings/jsc2/db/RoomRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 public interface RoomRepository extends JpaRepository<Room, Long> {
 
     @Query("SELECT id, floor, nro, type FROM Room r WHERE r.floor=?1 AND r.nro=?2")
-    Room findRoomByFloorAndNro(Long floor, Long nro);
+    Room findRoomByFloorAndNro(int floor, int nro);
 }
